requirements.txt
Django==5.0.1
djangorestframework==3.14.0
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
matplotlib==3.8.2
numpy==1.26.3
